Start enhanced support for flat files (ROM).

+++ ccdis.h
@@ -12,10 +12,10 @@
 #define GLOB_MC6502   0x2001
 #define GLOB_DATA     0x2002
 
-extern uint16_t mc_trace(const unsigned char *content, uint16_t size, int16_t *glob_index, uint16_t addr, int *new_labels);
-extern uint16_t cc_trace(const unsigned char *content, uint16_t size, int16_t *glob_index, uint16_t addr, int *new_labels);
-extern uint16_t mc_disassemble(FILE *ofp, const unsigned char *content, uint16_t size, int16_t *glob_index, uint16_t addr);
-extern uint16_t cc_disassemble(FILE *ofp, const unsigned char *content, uint16_t size, int16_t *glob_index, uint16_t addr);
+extern size_t mc_trace(const unsigned char *content, size_t size, int16_t *glob_index, size_t addr, int *new_labels);
+extern size_t cc_trace(const unsigned char *content, size_t size, int16_t *glob_index, size_t addr, int *new_labels);
+extern size_t mc_disassemble(FILE *ofp, const unsigned char *content, size_t size, int16_t *glob_index, size_t addr);
+extern size_t cc_disassemble(FILE *ofp, const unsigned char *content, size_t size, int16_t *glob_index, size_t addr);
 
 extern void print_label(FILE *ofp, int16_t *glob_index, uint16_t addr);
 extern uint16_t print_data(FILE *ofp, const unsigned char *content, uint16_t code_size, int16_t *glob_index, uint16_t addr);
